CORIN/requirements.txt
# Newer versions available

#Mysql database connection
cymysql
django-cymysql

#Restframework
djangorestframework
